refactor: modularize PerformanceScreen and isolate Vivo OEM packages
