+++ app/src/main/java/com/framex/app/ui/navigation/NavGraph.kt
@@ -10,10 +10,10 @@ import com.framex.app.ui.screens.AppearanceScreen
 import com.framex.app.ui.screens.DashboardScreen
 import com.framex.app.ui.screens.OnboardingScreen
 import com.framex.app.ui.screens.OverlayCustomizationScreen
-import com.framex.app.ui.screens.PerformanceScreen
 import com.framex.app.ui.screens.PermissionsScreen
 import com.framex.app.ui.screens.SplashScreen
 import com.framex.app.ui.screens.ThermalDiagnosticsScreen
+import com.framex.app.ui.screens.performance.PerformanceScreen
 
 sealed class Screen(val route: String) {
     object Splash : Screen("splash")
