+++ app/src/main/java/com/framex/app/ui/screens/performance/components/EsportsStatusRow.kt
@@ -1,0 +1,24 @@
+package com.framex.app.ui.screens.performance.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun EsportsStatusRow(label: String, value: String) {
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(label, color = Color.Gray, fontSize = 12.sp, fontWeight = FontWeight.Medium)
+        Text(value, color = Color.White, fontSize = 12.sp, fontWeight = FontWeight.Bold)
+    }
+}
